Changed tests
