+++ app/src/androidTest/java/ch/epfl/sweng/SDP/game/VotingPageActivityTest.java
@@ -1,5 +1,6 @@
 package ch.epfl.sweng.SDP.game;
 
+import static android.os.SystemClock.sleep;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
@@ -10,9 +11,11 @@ import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static ch.epfl.sweng.SDP.game.VotingPageActivity.disableAnimations;
 import static org.hamcrest.Matchers.is;
 
+import android.graphics.Canvas;
 import android.os.SystemClock;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
+import android.view.View;
 import android.widget.RatingBar;
 import ch.epfl.sweng.SDP.R;
 import org.junit.Rule;
@@ -58,5 +61,23 @@ public class VotingPageActivityTest {
     @Test
     public void rankingButtonIsDisplayed() {
         onView(withId(R.id.rankingButton)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void addStarsHandlesBigNumber() {
+        StarAnimationView starsAnimation =  mActivityRule.getActivity()
+                .findViewById(R.id.starsAnimation);
+        starsAnimation.onDraw(new Canvas());
+        starsAnimation.addStars(1000);
+        assert (5 == starsAnimation.getNumStars());
+    }
+
+    @Test
+    public void addStarsHandlesNegativeNumber() {
+        StarAnimationView starsAnimation =  mActivityRule.getActivity()
+                .findViewById(R.id.starsAnimation);
+        starsAnimation.onDraw(new Canvas());
+        starsAnimation.addStars(-10);
+        assert (0 == starsAnimation.getNumStars());
     }
 }
